refactor: move youtube URL calcs to transform
