+++ src/~utils.ts
@@ -50,24 +50,3 @@ export function stripHtmlHeadings(rawHtml: string): string {
   }
   return document.toString();
 }
-
-export function youtubeVidToEmbed(videoUrl: string): string {
-  return (
-    videoUrl
-      .replace("/watch?v=", "/embed/")
-      .replace("/shorts/", "/embed/")
-      .replace("https://www.youtube.com", "https://www.youtube-nocookie.com")
-      // map non-www URLs to www. Will raise SSL cert error without it!
-      // ex. https://nocookie-youtube.com/embed/HaFORudBWhQ
-      .replace("https://youtube.com", "https://www.youtube-nocookie.com") +
-    "?autoplay=1"
-  );
-}
-
-export function youtubeVidToThumbnail(shortUrl: string): string | undefined {
-  const idMatch = shortUrl.match(/[\/=]((\w|-)+)$/);
-  if (!idMatch) return undefined;
-
-  const [, id] = idMatch;
-  return `https://img.youtube.com/vi/${id}/hqdefault.jpg`;
-}
